requirements.txt
# Core dependencies
pydantic==2.11.7
playwright==1.55.0
google-search-results==2.4.2
transformers==4.56.1
torch==2.8.0
requests==2.32.5
python-dotenv==1.1.1
cohere==5.17.0

# API and deployment dependencies
fastapi==0.115.6
uvicorn[standard]==0.35.0
gunicorn==23.0.0
